test: split napi tests into separate files (#21475)
