+++ test/napi/node-napi-tests/test/js-native-api/3_callbacks/do.test.ts
@@ -1,0 +1,12 @@
+import { basename, dirname, sep } from "node:path";
+import { build, run } from "../../../harness";
+
+test("build", async () => {
+  await build(import.meta.dir);
+});
+
+for (const file of Array.from(new Bun.Glob("*.js").scanSync(import.meta.dir))) {
+  test(file, () => {
+    run(dirname(import.meta.dir), basename(import.meta.dir) + sep + file);
+  });
+}
