+++ test/napi/node-napi-tests/test/node-api/test_threadsafe_function/do.test.ts
@@ -1,0 +1,12 @@
+import { basename, dirname, sep } from "node:path";
+import { build, run } from "../../../harness";
+
+test("build", async () => {
+  await build(import.meta.dir);
+});
+
+for (const file of Array.from(new Bun.Glob("*.js").scanSync(import.meta.dir))) {
+  test.todoIf(["test.js", "test_legacy_uncaught_exception.js"].includes(file))(file, () => {
+    run(dirname(import.meta.dir), basename(import.meta.dir) + sep + file);
+  });
+}
