+++ test/no-validate-exceptions.txt
@@ -252,9 +252,29 @@ test/js/third_party/prisma/prisma.test.ts
 test/js/third_party/remix/remix.test.ts
 test/js/third_party/resvg/bbox.test.js
 test/js/third_party/rollup-v4/rollup-v4.test.ts
-test/napi/node-napi.test.ts
 test/napi/uv.test.ts
 test/napi/uv_stub.test.ts
+test/napi/node-napi-tests/test/node-api/test_fatal_exception/do.test.ts
+test/napi/node-napi-tests/test/js-native-api/test_typedarray/do.test.ts
+test/napi/node-napi-tests/test/node-api/test_async/do.test.ts
+test/napi/node-napi-tests/test/js-native-api/test_symbol/do.test.ts
+test/napi/node-napi-tests/test/node-api/1_hello_world/do.test.ts
+test/napi/node-napi-tests/test/js-native-api/test_instance_data/do.test.ts
+test/napi/node-napi-tests/test/js-native-api/test_object/do.test.ts
+test/napi/node-napi-tests/test/js-native-api/test_number/do.test.ts
+test/napi/node-napi-tests/test/js-native-api/test_new_target/do.test.ts
+test/napi/node-napi-tests/test/js-native-api/test_constructor/do.test.ts
+test/napi/node-napi-tests/test/js-native-api/test_handle_scope/do.test.ts
+test/napi/node-napi-tests/test/js-native-api/test_conversions/do.test.ts
+test/napi/node-napi-tests/test/js-native-api/test_dataview/do.test.ts
+test/napi/node-napi-tests/test/js-native-api/test_array/do.test.ts
+test/napi/node-napi-tests/test/js-native-api/test_date/do.test.ts
+test/napi/node-napi-tests/test/js-native-api/test_bigint/do.test.ts
+test/napi/node-napi-tests/test/js-native-api/test_cannot_run_js/do.test.ts
+test/napi/node-napi-tests/test/js-native-api/3_callbacks/do.test.ts
+test/napi/node-napi-tests/test/js-native-api/4_object_factory/do.test.ts
+test/napi/node-napi-tests/test/js-native-api/5_function_factory/do.test.ts
+test/napi/node-napi-tests/test/js-native-api/2_function_arguments/do.test.ts
 
 # normalizeCryptoAlgorithmParameters
 test/js/node/test/parallel/test-webcrypto-derivekey.js
